Fix final score calculation to use float
Updated the calculation of the final score to use float for more accurate results.

+++ main.py
@@ -30,7 +30,7 @@ def grade_calculator(e):
     units = (5, 3, 2) # number of hours per subject
 
     # average weighted score
-    final = int(((subjects[0] * units[0]) + (subjects[1] * units[0]) + (subjects[2] * units[1]) + (subjects[3] * units[0]) + (subjects[4] * units[1]) + (subjects[5] * units[2])) / (units[0] + units[0] + units[1] + units[0] + units[1] + units[2]))
+    final = float(((subjects[0] * units[0]) + (subjects[1] * units[0]) + (subjects[2] * units[1]) + (subjects[3] * units[0]) + (subjects[4] * units[1]) + (subjects[5] * units[2])) / (units[0] + units[0] + units[1] + units[0] + units[1] + units[2]))
 
     if final > 74:
         display(f'Weighted Average: {final}', target='output')
@@ -38,6 +38,7 @@ def grade_calculator(e):
     else:
         display(f'Weighted Average: {final}', target='output')
         display(f'You failed.', target='output')
+
